feat(tjs): rework course completion checking logic
